+++ adapters/swi.pl
@@ -4,7 +4,7 @@
 %  Copyright (c) 1998-2013 Paulo Moura <pmoura@logtalk.org>
 %
 %  Adapter file for SWI Prolog 6.0.0 and later versions
-%  Last updated on February 22, 2013
+%  Last updated on February 27, 2013
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -150,11 +150,11 @@ message_hook(discontiguous(_), _, _) :-			% SWI-Prolog discontiguous predicate
 '$lgt_prolog_meta_predicate'(use_module(_, _), use_module(*, *), predicate).
 
 
-% '$lgt_prolog_meta_directive'(@callable)
+% '$lgt_prolog_meta_directive'(@callable, @callable)
 
-'$lgt_prolog_meta_directive'(at_halt(0)).
-'$lgt_prolog_meta_directive'(initialization(0, *)).
-'$lgt_prolog_meta_directive'(thread_initialization(0)).
+'$lgt_prolog_meta_directive'(at_halt(_), at_halt(0)).
+'$lgt_prolog_meta_directive'(initialization(_, _), initialization(0, *)).
+'$lgt_prolog_meta_directive'(thread_initialization(_), thread_initialization(0)).